etl/capa_bronze/requirements.txt
pandas
requests
pyarrow
s3fs
boto3
polars
deltalake
duckdb
python-dotenv
requests_ntlm
